Move integrity check runnable into a separate file
